requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.4
deep-translator==1.11.4
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pyaes==1.6.1
pyasn1==0.6.2
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
soupsieve==2.8.3
Telethon==1.42.0
typing_extensions==4.15.0
urllib3==2.6.3
